Add some JavaDoc
